file upload and download, and pdf gen for invoices

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
@@ -17,9 +19,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) =>{
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+
+    if(!image) {
+        return res.status(422).render('admin/edit-product', {
+            path: 'admin/edit-product',
+            pageTitle: 'Add Product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Please attach a valid image',
+            validationErrors: []
+        });
+    }
 
     const errors = validationResult(req);
     if(!errors.isEmpty()) {
@@ -30,7 +48,6 @@ exports.postAddProduct = (req, res, next) =>{
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -38,6 +55,8 @@ exports.postAddProduct = (req, res, next) =>{
             validationErrors: errors.array()
         });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
@@ -106,7 +125,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
 
@@ -120,7 +139,6 @@ exports.postEditProduct = (req, res, next) => {
             product: {
                 _id: prodId,
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription
             },
@@ -135,7 +153,10 @@ exports.postEditProduct = (req, res, next) => {
                 return res.redirect('/');
             }
             product.title = updatedTitle;
-            product.imageUrl = updatedImageUrl;
+            if(image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             product.price = updatedPrice;
             product.description = updatedDescription;
 
@@ -155,7 +176,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId)
+        .then(product => {
+            if(!product){
+                next(new Error('Product not found'))
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: prodId, userId: req.user._id})
+        })
         .then(() => {
             res.redirect('/admin/products');
         })
